backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
PyMySQL==1.1.0
requests==2.31.0
argon2-cffi==23.1.0
pandas==2.2.2
pyarrow==17.0.0
cryptography==42.0.8